LayoutTests/platform/gtk/media/modern-media-controls/media-controller/media-controller-auto-hide-pause-expected.txt

Testing the controls bar remains visible when the media is paused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Video started playing, controls bar is visible by default.
FAIL mediaControls.classList.contains('faded') should be false. Was true.

Auto-hide timer would have elapsed, but media was paused so controls bar should remain visible.
PASS mediaControls.classList.contains('faded') is false

PASS successfullyParsed is true

TEST COMPLETE
